Route/Switch adjustment + new folder structure
Moved the Switch and Route elements to the bottom of the App file/removed the data mapping in the same file;  established new folder structure
