+++ src/recipe_components/Lunch.js
@@ -1,6 +1,6 @@
 import React from "react";
 import RecipeStyle from "./Recipestyle";
-import RecipesData1 from "./RecipeData1";
+import RecipesData1 from "../database/RecipeData1";
 
 export default function Lunch() {
   const recipeItems1 = RecipesData1.map((element, index) => {
