+++ src/recipe_components/Recipestyle.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./styles.css";
+import "../styles.css";
 
 export default function RecipeStyle(props) {
   const { title, recipeUrl, imgSrc, description } = props;
