+++ src/styles.css
@@ -30,10 +30,10 @@ a {
   min-height: 650px;
   margin: 10px;
   display: flex;
-  box-shadow: 1px 2px 3px rgba(0,0,0,0.2);
+  box-shadow: 1px 2px 3px rgba(0, 0, 0, 0.2);
 }
 
-.card img{
+.card img {
   height: 300px;
   object-fit: cover;
 }
@@ -42,7 +42,7 @@ a {
   display: flex;
   justify-content: center;
   margin-left: 10px;
-  margin-top:10px;
+  margin-top: 10px;
 }
 
 .layout-wrapper {
